Move request body serialization into ServiceClient.sendOperationRequest()
