+++ package.json
@@ -5,7 +5,7 @@
     "email": "azsdkteam@microsoft.com",
     "url": "https://github.com/Azure/ms-rest-js"
   },
-  "version": "0.13.0",
+  "version": "0.14.0",
   "description": "Isomorphic client Runtime for Typescript/node.js/browser javascript client libraries generated using AutoRest",
   "tags": [
     "isomorphic",
@@ -25,7 +25,7 @@
     "autorest",
     "clientruntime"
   ],
-  "main": "./dist/lib/msRest.js",
+  "main": "./lib/msRest.ts",
   "types": "./typings/lib/msRest.d.ts",
   "browser": {
     "./dist/lib/msRest.js": "./es/lib/msRest.js",
